refactor(ens160): expose raw resistance and baseline

+++ components/ens160/ens160.h
@@ -205,7 +205,7 @@ esp_err_t ens160_set_environmental_data(i2c_dev_t *dev, float temperature, float
  */
 esp_err_t
 ens160_measure(i2c_dev_t *dev, bool wait_for_new, ens160_aqi_t *aqi, uint16_t *tvoc, uint16_t *eco2,
-               uint32_t (*resistance)[4], uint32_t (*baseline)[4]);
+               uint16_t (*resistance)[4], uint16_t (*baseline)[4]);
 
 /**
  * @brief Reset baseline
